feat: adopt in-scope roadmap baseline (WorkRuntime progress, processing split, remote-source decomposition)
WorkRuntime: apply progress events to operation snapshots, progress-listener
plumbing through ProcessingContext and ProgressEmitter
Processing: extract run_dispatch, run_job, run_validation child modules;
remove dead wrapper stubs and fix cohesion imports
Remote-source: extract session_lifecycle from mod.rs; extract acquire-dialog
helpers into owned module
Work Center: remove legacy processing-progress overlay; model/state tests
updated to snapshot-only truth
Docs: api-map reflects Work Center uses snapshots as sole progress source

Focused evidence: 19 Rust tests pass (11 WorkRuntime + 8 processing/run),
5 Vitest tests pass (2 model + 3 state), cargo check clean (0 warnings).

This is an in-mission baseline; assertion completion remains assigned to
later feature workers (workruntime-lifecycle-closure,
remote-source-ui-decomposition, processing-run-cohesion-split).
