+++ docs/api-map.md
@@ -147,9 +147,8 @@ Source files:
 
 The `work-operation-*` events carry backend-authored operation snapshots
 (`src-tauri/src/work_runtime/`). Work Center is the background-operation
-consumer; it still overlays legacy `processing-progress` for fine-grained
-progress until that reducer is retired (see
-`docs/specs/pattern-cleanup-roadmap.md`).
+consumer and uses snapshot events as the sole progress source for accepted
+operations.
 
 Lifecycle event payloads carry `operation_kind` so Status Panel can distinguish
 merge processing, batch processing, and metadata save from backend truth instead
